# pizza_1337_preact/api_scanner_typescript/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/v1/pizza-types/{pizza_type_id}/dough": {
            "get": {
                "tags": [
                    "pizza_type"
                ],
                "summary": "Get Pizza Type Dough",
                "operationId": "get_pizza_type_dough_v1_pizza_types__pizza_type_id__dough_get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Pizza Type Id",
                            "type": "string",
                            "format": "uuid"
                        },
                        "name": "pizza_type_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DoughSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/doughs": {
            "get": {
                "tags": [
                    "dough"
                ],
                "summary": "Get All Doughs",
                "operationId": "get_all_doughs_v1_doughs_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Get All Doughs V1 Doughs Get",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DoughListItemSchema"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "dough"
                ],
                "summary": "Create Dough",
                "operationId": "create_dough_v1_doughs_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DoughCreateSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DoughSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/doughs/{dough_id}": {
            "get": {
                "tags": [
                    "dough"
                ],
                "summary": "Get Dough",
                "operationId": "get_dough_v1_doughs__dough_id__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Dough Id",
                            "type": "string",
                            "format": "uuid"
                        },
                        "name": "dough_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DoughSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "dough"
                ],
                "summary": "Update Dough",
                "operationId": "update_dough_v1_doughs__dough_id__put",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Dough Id",
                            "type": "string",
                            "format": "uuid"
                        },
                        "name": "dough_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DoughCreateSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DoughSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "dough"
                ],
                "summary": "Delete Dough",
                "operationId": "delete_dough_v1_doughs__dough_id__delete",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Dough Id",
                            "type": "string",
                            "format": "uuid"
                        },
                        "name": "dough_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DoughCreateSchema": {
                "title": "DoughCreateSchema",
                "required": [
                    "name",
                    "price",
                    "description",
                    "stock"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "price": {
                        "title": "Price",
                        "type": "number"
                    },
                    "description": {
                        "title": "Description",
                        "type": "string"
                    },
                    "stock": {
                        "title": "Stock",
                        "type": "integer"
                    }
                }
            },
            "DoughListItemSchema": {
                "title": "DoughListItemSchema",
                "required": [
                    "id",
                    "name",
                    "price",
                    "description"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "price": {
                        "title": "Price",
                        "type": "number"
                    },
                    "description": {
                        "title": "Description",
                        "type": "string"
                    }
                }
            },
            "DoughSchema": {
                "title": "DoughSchema",
                "required": [
                    "name",
                    "price",
                    "description",
                    "stock",
                    "id"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "price": {
                        "title": "Price",
                        "type": "number"
                    },
                    "description": {
                        "title": "Description",
                        "type": "string"
                    },
                    "stock": {
                        "title": "Stock",
                        "type": "integer"
                    },
                    "id": {
                        "title": "Id",
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
...
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            }
        }
    },
...
}