# pizza_1337_preact/api_scanner_typescript/src/api/models/DoughSchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DoughSchema = {
    name: string;
    price: number;
    description: string;
    stock: number;
    id: string;
};